+++ test/unit/fuzz-tools.test.js
@@ -380,6 +380,7 @@ import * as nyhusHerniaV400 from '../../lib/nyhus-hernia-v400.js';
 import * as zargarCausticV401 from '../../lib/zargar-caustic-v401.js';
 import * as laugeHansenV402 from '../../lib/lauge-hansen-v402.js';
 import * as berndtHartyV403 from '../../lib/berndt-harty-v403.js';
+import * as reganMorreyV404 from '../../lib/regan-morrey-v404.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -705,6 +706,7 @@ const MODULES = {
   'zargar-caustic-v401.js': zargarCausticV401,
   'lauge-hansen-v402.js': laugeHansenV402,
   'berndt-harty-v403.js': berndtHartyV403,
+  'regan-morrey-v404.js': reganMorreyV404,
   'tb-testing.js': tbTesting,
 };